moved getCompletedTasksByVolunteer from Tasks.js to new file called volunteerProfile.js, with new functions inside too
