+++ functions/index.js
@@ -7,3 +7,4 @@ exports.checkAccountType=require("./checkAccountType");
 exports.stickers = require("./stickers");
 exports.chats = require("./chats");
 exports.advice = require("./advice");
+exports.volunteerProfile = require("./volunteerProfile");
